+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Serverless Plugin Neo Changelog
 
+## 0.1.1 - April 20, 2022
+
+- Add handling for specific dependencies that can't be traced fully
+- Improve package size display
+- Update dependencies
+
 ## 0.1.0 - April 18, 2022
 
 - Initial release